Moved all created QuizGroup methods to quiz.py and quiz_group.py. Moved related test methods and fixtures as well.
