+++ tests/fixtures/course.json
@@ -1392,59 +1392,6 @@
 		],
 		"status_code": 200
 	},
-	"get_quiz_group": {
-		"method": "GET",
-		"endpoint": "courses/1/quizzes/1/groups/1",
-		"data": {
-			"id": 1,
-			"quiz_id": 1,
-			"name": "Test Group",
-			"pick_count": 1,
-			"question_points": 2,
-			"assessment_question_bank_id": 3
-		},
-		"status_code": 200
-	},
-	"create_question_group": {
-		"method": "POST",
-		"endpoint": "courses/1/quizzes/1/groups",
-		"data": {"quiz_groups": [
-			{
-				"id": 1,
-				"quiz_id": 1,
-				"name": "Test Group",
-				"pick_count": 1,
-				"question_points": 2,
-				"assessment_question_bank_id": 3
-			}
-		]},
-		"status_code": 200
-	},
-	"update_question_group": {
-		"method": "PUT",
-		"endpoint": "courses/1/quizzes/1/groups/1",
-		"data": {"quiz_groups": [
-			{
-				"id": 1,
-				"quiz_id": 1,
-				"name": "Test Group",
-				"pick_count": 1,
-				"question_points": 2,
-				"assessment_question_bank_id": 3
-			}
-		]},
-		"status_code": 200
-	},
-	"delete_question_group": {
-		"method": "DELETE",
-		"endpoint": "courses/1/quizzes/1/groups/1",
-		"status_code": 204
-	},
-	"reorder_question_group": {
-		"method": "POST",
-		"endpoint": "courses/1/quizzes/1/groups/1/reorder",
-		"status_code": 204
-	},
 	"add_grading_standards": {
 		"method": "POST",
 		"endpoint": "courses/1/grading_standards",
